Largescale changes for NAG compilation.
THIS CODE WILL CURRENTLY NOT WORK OR LINK.
Key routine calls are commented out to
check compilation elsewhere.
These are:
FciMCPar: assign_proc call  [function pointer issue]
Calc.F90: Two calls to calcrhopii2 [ajwts hidden variables issues]
error_handling_neci.F90 stop_all_c [c interface issues]

These need fixing!

+++ src/CCMC.F90
@@ -2758,9 +2758,9 @@ SUBROUTINE ReadPopsFileCCMC(DetList,nMaxAmpl,nAmpl,dNorm)
       INTEGER(kind=n_int), pointer :: DetList(:,:)
       logical :: tPop64Bit,tPopHPHF,tPopLz
       integer :: iPopLenof_sign,iPopNel,iPopIter,PopNIfD,PopNIfY,PopNIfSgn,PopNIfFlag,PopNIfTot
-      integer(8) :: iPopAllTotWalkers
-      real(8) :: PopDiagSft
-      integer(8) , dimension(lenof_sign) :: PopSumNoatHF
+      integer(int64) :: iPopAllTotWalkers
+      real(dp) :: PopDiagSft
+      integer(int64) , dimension(lenof_sign) :: PopSumNoatHF
       integer nMaxAmpl,nAmpl
       integer(int64) , dimension(lenof_sign) :: CurrParts
       integer :: ReadBatch    !This parameter determines the length of the array to batch read in walkers from a popsfile
